requirements.txt
Django~=3.2.7
django-debug-toolbar~=3.2.2
environs~=9.3.3
djangorestframework~=3.12.4
